pyproject.toml
[build-system]
requires = ["uv", "setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pre-commit-template"
version = "0.1.1"
description = "Auto-detect technologies and generate pre-commit configuration files."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jakub Kriz"}
]
dependencies = [
    "PyYAML>=6.0.0",
    "rich>=13.0.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "mypy>=1.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
pre-commit-template = "pre_commit_template.main:main"

[tool.setuptools]
packages = ["pre_commit_template", "pre_commit_template.hook_templates"]

[tool.setuptools.package-data]
"*" = ["py.typed"]
"pre_commit_template" = ["hook_templates/*.j2"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
]
